requirements.txt
# Python 3.11+
# Install with: pip install -r requirements.txt

# Web framework
fastapi==0.136.0
uvicorn==0.44.0

# HTTP
requests==2.33.1
httpx==0.28.1

# Data
numpy==2.1.3
pandas==2.2.3
scipy==1.14.1

# MongoDB (async)
motor==3.6.0
pymongo>=4.9,<4.10

# Gemini
google-genai==1.73.1

# Env
python-dotenv==1.0.1

# Testing
pytest>=8.2,<9
pytest-asyncio==0.24.0